==> notebook/00-Data_Gather.ipynb <==
# Data Gather from Kaggle

from dotenv import load_dotenv, find_dotenv
import os
import sys 

_ = load_dotenv(find_dotenv())


kaggle_usrname = os.environ["KAGGLE_USERNAME"]
kaggle_key = os.environ["KAGGLE_KEY"]
project_path = os.environ["PROJECT_ABS_PATH"]

sys.path.append(os.path.abspath(os.path.join(project_path)))


from src.data_acess.DataRetrieverFactory import DataRetrieverFactory
from src.data_acess.KaggleRetrieverStrategy import KaggleRetrieverStrategy


dataset_name = "fedesoriano/heart-failure-prediction"
save_path = r"""../data/"""


data_retriever = DataRetrieverFactory(KaggleRetrieverStrategy())
data_retriever.get_data(dataset_name = dataset_name, 
                        save_path = save_path)
