==> notebook/02-Data_Clean_Transformation.ipynb <==
# Data Transformation and Clean

#### Read the downloaded data

from dotenv import load_dotenv, find_dotenv
import os
import sys 

_ = load_dotenv(find_dotenv())

project_path = os.environ["PROJECT_ABS_PATH"]
sys.path.append(os.path.abspath(os.path.join(project_path)))


from sklearn.preprocessing import StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder
import pandas as pd
import numpy as np

from src.sampling.SplitData import SplitData

#### Get the data from the previous step

heart_df = pd.read_csv('../data/heart.csv', sep = ',')
heart_df.rename(columns={'HeartDisease':'label'}, inplace=True)


heart_df.head(5)

#### Split into train, dev and test

splitter = SplitData(partitions = [0.8, 0.2])


heart_df_label = heart_df.pop('label')

train, train_label, test, test_label = \
                                    splitter.get_two_sets(heart_df, 
                                                          heart_df_label)

print(train.shape)
print(test.shape)

#### Perform the transformation over the numerical and categorical variable

std_scaler_obj = StandardScaler()
ohe_obj = OneHotEncoder(sparse_output = False, 
                        handle_unknown = 'ignore')

# Lista das colunas numericas
train_num = train.select_dtypes(include=[np.number])
num_cols = train_num.columns

# Lista das colunas categóricas
train_cat = train.select_dtypes(include=['object'])
cat_cols = train_cat.columns


# Criar o transformador
preprocessor = ColumnTransformer(
    transformers = [
                     ('stdscaler', std_scaler_obj, num_cols) 
                    ,('ohe', ohe_obj, cat_cols)
                    ],
    remainder = 'drop'  # Mantém as colunas não transformadas
)


train_transf_output = pd.DataFrame(
                             preprocessor.fit_transform(train) 
                            ,columns = preprocessor.get_feature_names_out()
                            ,index=train.index
                         )

test_transf_output = pd.DataFrame(
                             preprocessor.transform(test) 
                            ,columns = preprocessor.get_feature_names_out()
                            ,index=test.index
                         )


print(train_transf_output.columns)
print(train_transf_output.shape)

print(test_transf_output.columns)
print(test_transf_output.shape)

#### Save the transformed dataset predictors

train_transf_output.to_csv('../data/train_transformed.csv', sep = ';', index = True)
test_transf_output.to_csv('../data/test_transformed.csv', sep = ';', index = True)

#### Save the transformed dataset labels

train_label_df = pd.DataFrame(
                             train_label
                            ,index=train_label.index
                         )

test_label_df = pd.DataFrame(
                             test_label
                            ,index=test_label.index
                         )



train_label_df.to_csv('../data/train_label_df.csv', sep = ';', index = True)
test_label_df.to_csv('../data/test_label_df.csv', sep = ';', index = True)
