==> notebook/03-Feature_Engineering.ipynb <==
# Feature Engineering

#### Read the transformed data

import pandas as pd

train_predictors = pd.read_csv('../data/train_transformed.csv', sep = ';', index_col=0)
test_predictors = pd.read_csv('../data/test_transformed.csv', sep = ';', index_col=0)
train_label = pd.read_csv('../data/train_label_df.csv', sep = ';', index_col=0)
test_label = pd.read_csv('../data/test_label_df.csv', sep = ';', index_col=0)


#### save the datasets with the feature engineered

train_predictors.to_csv('../data/train_predictors_FE.csv', sep = ';', index = True)
test_predictors.to_csv('../data/test_predictors_FE.csv', sep = ';', index = True)
train_label.to_csv('../data/train_label_FE.csv', sep = ';', index = True)
test_label.to_csv('../data/test_label_FE.csv', sep = ';', index = True)
